Preprocessor variables to int8_t: save space

+++ src/ALog.cpp
@@ -67,29 +67,29 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // True for all: UNO (ATMega168/328)
 #if defined(__AVR_ATmega168__) || defined(__AVR_ATmega328P__) || defined(__AVR_ATmega8__) || defined(__AVR_ATmega88__)
   // SD card: CSpin and protected pins
-  const int SCKpin = 13;
-  const int MISOpin = 12;
-  const int MOSIpin = 11;
-  const int CSpin = 10;
+  const int8_t SCKpin = 13;
+  const int8_t MISOpin = 12;
+  const int8_t MOSIpin = 11;
+  const int8_t CSpin = 10;
   // Protected I2C pins
-  const int SDApin = A4;
-  const int SCLpin = A5;
+  const int8_t SDApin = A4;
+  const int8_t SCLpin = A5;
   // Digital pins
-  const int SensorPowerPin = 4; // Activates voltage regulator to give power to sensors
+  const int8_t SensorPowerPin = 4; // Activates voltage regulator to give power to sensors
   // Sleep mode pins
-  const int wakePin = 2; // interrupt pin used for waking up via the alarm
-  const int manualWakePin = 5; // Wakes the logger with a manual button - overrides the "wait for right minute" commands
+  const int8_t wakePin = 2; // interrupt pin used for waking up via the alarm
+  const int8_t manualWakePin = 5; // Wakes the logger with a manual button - overrides the "wait for right minute" commands
   #if defined(ARDUINO_AVR_ALOG_BOTTLELOGGER_V2)
     // 7 for both? Maybe this is better than redefining variables.
     // GitHub issue is open on this; should guide future decisions
-    int SDpowerPin = 7; // Turns on voltage source to SD card
-    int RTCpowerPin = 7; // Activates voltage regulator to power the RTC (otherwise is on backup power from VCC or batt)
-    int LEDpin = 8; // LED to tell user if logger is working properly  
+    int8_t SDpowerPin = 7; // Turns on voltage source to SD card
+    int8_t RTCpowerPin = 7; // Activates voltage regulator to power the RTC (otherwise is on backup power from VCC or batt)
+    int8_t LEDpin = 8; // LED to tell user if logger is working properly  
   #elif defined(ARDUINO_AVR_ALOG_BOTTLELOGGER_PRE_V200)
     // IS IT EVEN TRUE HERE THAT THERE ARE MULTIPLE PINS USED FOR SD AND CLOCK? START HERE!
-    int SDpowerPin = 8; // Turns on voltage source to SD card
-    int RTCpowerPin = 6; // Activates voltage regulator to power the RTC (otherwise is on backup power from VCC or batt)
-    int LEDpin = 9; // LED to tell user if logger is working properly
+    int8_t SDpowerPin = 8; // Turns on voltage source to SD card
+    int8_t RTCpowerPin = 6; // Activates voltage regulator to power the RTC (otherwise is on backup power from VCC or batt)
+    int8_t LEDpin = 9; // LED to tell user if logger is working properly
   #else
     // These can be variable; default values here are for the ALog shield
     // No pin is appropriate for the ALog shield for SDpowerPin and RTCpowerPin,
@@ -102,15 +102,15 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // True for all: MEGA (ATMega1280/2560)
 #if defined(__AVR_ATmega1280__) || defined(__AVR_ATmega2560__)
   // SD card: CSpin and protected pins
-  const int SCKpin = 52;
-  const int MISOpin = 50;
-  const int MOSIpin = 51;
-  const int CSpin = 53;
+  const int8_t SCKpin = 52;
+  const int8_t MISOpin = 50;
+  const int8_t MOSIpin = 51;
+  const int8_t CSpin = 53;
   // Protected I2C pins
-  const int SDApin = 20;
-  const int SCLpin = 21;
+  const int8_t SDApin = 20;
+  const int8_t SCLpin = 21;
   // Sleep mode pins: same as for 328
-  const int wakePin = 2; // interrupt pin used for waking up via the alarm
+  const int8_t wakePin = 2; // interrupt pin used for waking up via the alarm
 #endif
 
 /*
